Support multiple story flows for an artifact
